chore(wdyr): move to utils
